app/requirements.txt
Flask>=3.0
google-generativeai>=0.3.0
